rate_np. should be a dropin replacement for rate, +returns sim stats
I haven't removed the rate function out of an abundance of caution,
but as far as I can tell rate_np and rate _should_ behave identically.

we currently only run sims for artifacts that have 4 substats already.
it's possible but significantly more annoying to simulate substat
generation rolls; also, how would you report it

performance is reasonable. compute takes on the order of 50ms, when
it used to take on the order of 0.5ms—i.e., should be indistinguishable
given network latency.